requirements.txt
pikepdf==9.3.0
Pillow==10.4.0
